Move displayscore.js to components folder
