+++ src/index.js
@@ -9,7 +9,7 @@ import {routerReducer} from 'react-router-redux';
 
 import NotFound from "./pages/404"
 import Homepage from './pages/Homepage';
-import DisplayScore from './pages/displayscore';
+import DisplayScore from './components/scoreboard/displayscore';
 import './css/main.scss';
 
 // we'll worry about redux later I just set this up so that way I can set up the redux router
